Added functionality to parameterize interaction for liquid/gas CP
One scalar and one vector interaction can be now optimized to fulfill saturation and liquid/gas critical point criteria.
A given range of coefficient and exponent ranges is searched in parallel and given to GSL optimization. In order to find
interactions which agree with experimental values, it is also possible to search a range of critical points. The interactions
found are validated by demanding a precision of reproducing the critical point.